Corrections concerning ClientInfo in .proto
client_id in ChatMessage has been replaced with ClientInfo so received messages can
display username instead of id.
Subscribe service has also been modified so it takes a ClientInfo instead of an Empty
object. This should enable the subscribe method to add the client id to the map of
clients in the server struct.