Removed constant broadcast message
Commented out contents of for loop in server's subscribe method. Seems to still work.
Kept the time.Sleep since it seems to keep the for true loop from using 100% of cpu.
Deleting the sleep as well had server using 50% of my cpu at one point.
Keeping it seems not to affect the program and makes cpu usage vanishingly small.